Move webxr-related code in canvas crate to webxr module
