+++ components/canvas/lib.rs
@@ -12,4 +12,6 @@ pub mod canvas_data;
 pub mod canvas_paint_thread;
 mod webgl_limits;
 mod webgl_mode;
+#[cfg(feature = "webxr")]
+mod webxr;
 pub mod webgl_thread;
