+++ components/canvas/webgl_thread.rs
@@ -3,8 +3,6 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 use std::borrow::Cow;
-#[cfg(feature = "webxr")]
-use std::collections::HashMap;
 use std::rc::Rc;
 use std::sync::{Arc, Mutex};
 use std::{slice, thread};
@@ -12,8 +10,9 @@ use std::{slice, thread};
 use bitflags::bitflags;
 use byteorder::{ByteOrder, NativeEndian, WriteBytesExt};
 use canvas_traits::webgl;
+use canvas_traits::webgl::WebXRCommand;
 use canvas_traits::webgl::{
-    webgl_channel, ActiveAttribInfo, ActiveUniformBlockInfo, ActiveUniformInfo, AlphaTreatment,
+    ActiveAttribInfo, ActiveUniformBlockInfo, ActiveUniformInfo, AlphaTreatment,
     GLContextAttributes, GLLimits, GlType, InternalFormatIntVec, ProgramLinkInfo, TexDataType,
     TexFormat, WebGLBufferId, WebGLChan, WebGLCommand, WebGLCommandBacktrace, WebGLContextId,
     WebGLCreateContextResult, WebGLFramebufferBindingRequest, WebGLFramebufferId, WebGLMsg,
@@ -21,8 +20,6 @@ use canvas_traits::webgl::{
     WebGLSLVersion, WebGLSamplerId, WebGLSender, WebGLShaderId, WebGLSyncId, WebGLTextureId,
     WebGLVersion, WebGLVertexArrayId, YAxisTreatment,
 };
-#[cfg(feature = "webxr")]
-use canvas_traits::webgl::{WebXRCommand, WebXRLayerManagerId};
 use euclid::default::Size2D;
 use fnv::FnvHashMap;
 use half::f16;
@@ -42,23 +39,14 @@ use webrender_api::{
     ImageData, ImageDescriptor, ImageDescriptorFlags, ImageFormat, ImageKey,
 };
 use webrender_traits::{WebrenderExternalImageRegistry, WebrenderImageHandlerType};
-#[cfg(feature = "webxr")]
-use webxr::SurfmanGL as WebXRSurfman;
-#[cfg(feature = "webxr")]
-use webxr_api::{
-    ContextId as WebXRContextId, Error as WebXRError, GLContexts as WebXRContexts,
-    GLTypes as WebXRTypes, LayerGrandManager as WebXRLayerGrandManager,
-    LayerGrandManagerAPI as WebXRLayerGrandManagerAPI, LayerId as WebXRLayerId,
-    LayerInit as WebXRLayerInit, LayerManager as WebXRLayerManager,
-    LayerManagerAPI as WebXRLayerManagerAPI, LayerManagerFactory as WebXRLayerManagerFactory,
-    SubImages as WebXRSubImages,
-};
+use webxr_api::ContextId as WebXRContextId;
 
 use crate::webgl_limits::GLLimitsDetect;
+use crate::webxr::{WebXRBridge, WebXRBridgeContexts, WebXRBridgeInit};
 
-struct GLContextData {
-    ctx: Context,
-    gl: Rc<Gl>,
+pub(crate) struct GLContextData {
+    pub(crate) ctx: Context,
+    pub(crate) gl: Rc<Gl>,
     state: GLState,
     attributes: GLContextAttributes,
 }
@@ -91,18 +79,18 @@ pub struct GLState {
 impl GLState {
     // Are we faking having no alpha / depth / stencil?
     fn fake_no_alpha(&self) -> bool {
-        self.drawing_to_default_framebuffer &
-            !self.requested_flags.contains(ContextAttributeFlags::ALPHA)
+        self.drawing_to_default_framebuffer
+            & !self.requested_flags.contains(ContextAttributeFlags::ALPHA)
     }
 
     fn fake_no_depth(&self) -> bool {
-        self.drawing_to_default_framebuffer &
-            !self.requested_flags.contains(ContextAttributeFlags::DEPTH)
+        self.drawing_to_default_framebuffer
+            & !self.requested_flags.contains(ContextAttributeFlags::DEPTH)
     }
 
     fn fake_no_stencil(&self) -> bool {
-        self.drawing_to_default_framebuffer &
-            !self
+        self.drawing_to_default_framebuffer
+            & !self
                 .requested_flags
                 .contains(ContextAttributeFlags::STENCIL)
     }
@@ -395,10 +383,7 @@ impl WebGLThread {
     /// Handles a WebXR message
     fn handle_webxr_command(&mut self, command: WebXRCommand) {
         trace!("processing {:?}", command);
-        let mut contexts = WebXRBridgeContexts {
-            contexts: &mut self.contexts,
-            bound_context_id: &mut self.bound_context_id,
-        };
+        let mut contexts = WebXRBridgeContexts::new(&mut self.contexts, &mut self.bound_context_id);
         match command {
             WebXRCommand::CreateLayerManager(sender) => {
                 let result = self
@@ -502,10 +487,10 @@ impl WebGLThread {
         // WebGL requires all contexts to be able to create framebuffers with
         // alpha, depth and stencil. So we always create a context with them,
         // and fake not having them if requested.
-        let flags = requested_flags |
-            ContextAttributeFlags::ALPHA |
-            ContextAttributeFlags::DEPTH |
-            ContextAttributeFlags::STENCIL;
+        let flags = requested_flags
+            | ContextAttributeFlags::ALPHA
+            | ContextAttributeFlags::DEPTH
+            | ContextAttributeFlags::STENCIL;
         let context_attributes = &ContextAttributes {
             version: webgl_version.to_surfman_version(self.api_type),
             flags,
@@ -727,10 +712,8 @@ impl WebGLThread {
         {
             // Destroy WebXR layers associated with this context
             let webxr_context_id = WebXRContextId::from(context_id);
-            let mut webxr_contexts = WebXRBridgeContexts {
-                contexts: &mut self.contexts,
-                bound_context_id: &mut self.bound_context_id,
-            };
+            let mut webxr_contexts =
+                WebXRBridgeContexts::new(&mut self.contexts, &mut self.bound_context_id);
             self.webxr_bridge.destroy_all_layers(
                 &mut self.device,
                 &mut webxr_contexts,
@@ -852,7 +835,7 @@ impl WebGLThread {
     }
 
     /// Gets a reference to a Context for a given WebGLContextId and makes it current if required.
-    fn make_current_if_needed<'a>(
+    pub(crate) fn make_current_if_needed<'a>(
         device: &Device,
         context_id: WebGLContextId,
         contexts: &'a FnvHashMap<WebGLContextId, GLContextData>,
@@ -871,7 +854,7 @@ impl WebGLThread {
     }
 
     /// Gets a mutable reference to a GLContextWrapper for a WebGLContextId and makes it current if required.
-    fn make_current_if_needed_mut<'a>(
+    pub(crate) fn make_current_if_needed_mut<'a>(
         device: &Device,
         context_id: WebGLContextId,
         contexts: &'a mut FnvHashMap<WebGLContextId, GLContextData>,
@@ -2613,10 +2596,10 @@ fn image_to_tex_image_data(
     }
 
     match (format, data_type) {
-        (TexFormat::RGBA, TexDataType::UnsignedByte) |
-        (TexFormat::RGBA8, TexDataType::UnsignedByte) => pixels,
-        (TexFormat::RGB, TexDataType::UnsignedByte) |
-        (TexFormat::RGB8, TexDataType::UnsignedByte) => {
+        (TexFormat::RGBA, TexDataType::UnsignedByte)
+        | (TexFormat::RGBA8, TexDataType::UnsignedByte) => pixels,
+        (TexFormat::RGB, TexDataType::UnsignedByte)
+        | (TexFormat::RGB8, TexDataType::UnsignedByte) => {
             for i in 0..pixel_count {
                 let rgb = {
                     let rgb = &pixels[i * 4..i * 4 + 3];
@@ -2659,10 +2642,10 @@ fn image_to_tex_image_data(
             for i in 0..pixel_count {
                 let p = {
                     let rgba = &pixels[i * 4..i * 4 + 4];
-                    (rgba[0] as u16 & 0xf0) << 8 |
-                        (rgba[1] as u16 & 0xf0) << 4 |
-                        (rgba[2] as u16 & 0xf0) |
-                        (rgba[3] as u16 & 0xf0) >> 4
+                    (rgba[0] as u16 & 0xf0) << 8
+                        | (rgba[1] as u16 & 0xf0) << 4
+                        | (rgba[2] as u16 & 0xf0)
+                        | (rgba[3] as u16 & 0xf0) >> 4
                 };
                 NativeEndian::write_u16(&mut pixels[i * 2..i * 2 + 2], p);
             }
@@ -2673,10 +2656,10 @@ fn image_to_tex_image_data(
             for i in 0..pixel_count {
                 let p = {
                     let rgba = &pixels[i * 4..i * 4 + 4];
-                    (rgba[0] as u16 & 0xf8) << 8 |
-                        (rgba[1] as u16 & 0xf8) << 3 |
-                        (rgba[2] as u16 & 0xf8) >> 2 |
-                        (rgba[3] as u16) >> 7
+                    (rgba[0] as u16 & 0xf8) << 8
+                        | (rgba[1] as u16 & 0xf8) << 3
+                        | (rgba[2] as u16 & 0xf8) >> 2
+                        | (rgba[3] as u16) >> 7
                 };
                 NativeEndian::write_u16(&mut pixels[i * 2..i * 2 + 2], p);
             }
@@ -2687,9 +2670,9 @@ fn image_to_tex_image_data(
             for i in 0..pixel_count {
                 let p = {
                     let rgb = &pixels[i * 4..i * 4 + 3];
-                    (rgb[0] as u16 & 0xf8) << 8 |
-                        (rgb[1] as u16 & 0xfc) << 3 |
-                        (rgb[2] as u16 & 0xf8) >> 3
+                    (rgb[0] as u16 & 0xf8) << 8
+                        | (rgb[1] as u16 & 0xfc) << 3
+                        | (rgb[2] as u16 & 0xf8) >> 3
                 };
                 NativeEndian::write_u16(&mut pixels[i * 2..i * 2 + 2], p);
             }
@@ -2725,8 +2708,8 @@ fn image_to_tex_image_data(
             pixels
         },
 
-        (TexFormat::Luminance, TexDataType::Float) |
-        (TexFormat::Luminance32f, TexDataType::Float) => {
+        (TexFormat::Luminance, TexDataType::Float)
+        | (TexFormat::Luminance32f, TexDataType::Float) => {
             for rgba8 in pixels.chunks_mut(4) {
                 let p = rgba8[0] as f32;
                 NativeEndian::write_f32(rgba8, p);
@@ -2734,8 +2717,8 @@ fn image_to_tex_image_data(
             pixels
         },
 
-        (TexFormat::LuminanceAlpha, TexDataType::Float) |
-        (TexFormat::LuminanceAlpha32f, TexDataType::Float) => {
+        (TexFormat::LuminanceAlpha, TexDataType::Float)
+        | (TexFormat::LuminanceAlpha32f, TexDataType::Float) => {
             let mut data = Vec::<u8>::with_capacity(pixel_count * 8);
             for rgba8 in pixels.chunks(4) {
                 data.write_f32::<NativeEndian>(rgba8[0] as f32).unwrap();
@@ -2744,8 +2727,8 @@ fn image_to_tex_image_data(
             data
         },
 
-        (TexFormat::RGBA, TexDataType::HalfFloat) |
-        (TexFormat::RGBA16f, TexDataType::HalfFloat) => {
+        (TexFormat::RGBA, TexDataType::HalfFloat)
+        | (TexFormat::RGBA16f, TexDataType::HalfFloat) => {
             let mut rgbaf16 = Vec::<u8>::with_capacity(pixel_count * 8);
             for rgba8 in pixels.chunks(4) {
                 rgbaf16
@@ -2779,8 +2762,8 @@ fn image_to_tex_image_data(
             }
             rgbf16
         },
-        (TexFormat::Alpha, TexDataType::HalfFloat) |
-        (TexFormat::Alpha16f, TexDataType::HalfFloat) => {
+        (TexFormat::Alpha, TexDataType::HalfFloat)
+        | (TexFormat::Alpha16f, TexDataType::HalfFloat) => {
             for i in 0..pixel_count {
                 let p = f16::from_f32(pixels[i * 4 + 3] as f32).to_bits();
                 NativeEndian::write_u16(&mut pixels[i * 2..i * 2 + 2], p);
@@ -2788,8 +2771,8 @@ fn image_to_tex_image_data(
             pixels.truncate(pixel_count * 2);
             pixels
         },
-        (TexFormat::Luminance, TexDataType::HalfFloat) |
-        (TexFormat::Luminance16f, TexDataType::HalfFloat) => {
+        (TexFormat::Luminance, TexDataType::HalfFloat)
+        | (TexFormat::Luminance16f, TexDataType::HalfFloat) => {
             for i in 0..pixel_count {
                 let p = f16::from_f32(pixels[i * 4] as f32).to_bits();
                 NativeEndian::write_u16(&mut pixels[i * 2..i * 2 + 2], p);
@@ -2797,8 +2780,8 @@ fn image_to_tex_image_data(
             pixels.truncate(pixel_count * 2);
             pixels
         },
-        (TexFormat::LuminanceAlpha, TexDataType::HalfFloat) |
-        (TexFormat::LuminanceAlpha16f, TexDataType::HalfFloat) => {
+        (TexFormat::LuminanceAlpha, TexDataType::HalfFloat)
+        | (TexFormat::LuminanceAlpha16f, TexDataType::HalfFloat) => {
             for rgba8 in pixels.chunks_mut(4) {
                 let lum = f16::from_f32(rgba8[0] as f32).to_bits();
                 let a = f16::from_f32(rgba8[3] as f32).to_bits();
@@ -2842,10 +2825,10 @@ fn premultiply_inplace(format: TexFormat, data_type: TexDataType, pixels: &mut [
                 let a = extend_to_8_bits(pix & 0x0f);
                 NativeEndian::write_u16(
                     rgba,
-                    ((pixels::multiply_u8_color(r, a) & 0xf0) as u16) << 8 |
-                        ((pixels::multiply_u8_color(g, a) & 0xf0) as u16) << 4 |
-                        ((pixels::multiply_u8_color(b, a) & 0xf0) as u16) |
-                        ((a & 0x0f) as u16),
+                    ((pixels::multiply_u8_color(r, a) & 0xf0) as u16) << 8
+                        | ((pixels::multiply_u8_color(g, a) & 0xf0) as u16) << 4
+                        | ((pixels::multiply_u8_color(b, a) & 0xf0) as u16)
+                        | ((a & 0x0f) as u16),
                 );
             }
         },
@@ -2872,8 +2855,8 @@ fn flip_pixels_y(
     unpacking_alignment: usize,
     pixels: Vec<u8>,
 ) -> Vec<u8> {
-    let cpp = (data_type.element_size() * internal_format.components() /
-        data_type.components_per_element()) as usize;
+    let cpp = (data_type.element_size() * internal_format.components()
+        / data_type.components_per_element()) as usize;
 
     let stride = (width * cpp + unpacking_alignment - 1) & !(unpacking_alignment - 1);
 
@@ -3029,324 +3012,324 @@ impl FramebufferRebindingInfo {
     }
 }
 
-#[cfg(feature = "webxr")]
-/// Bridge between WebGL and WebXR
-pub(crate) struct WebXRBridge {
-    factory_receiver: crossbeam_channel::Receiver<WebXRLayerManagerFactory<WebXRSurfman>>,
-    managers: HashMap<WebXRLayerManagerId, Box<dyn WebXRLayerManagerAPI<WebXRSurfman>>>,
-    next_manager_id: u32,
-}
+// #[cfg(feature = "webxr")]
+// /// Bridge between WebGL and WebXR
+// pub(crate) struct WebXRBridge {
+//     factory_receiver: crossbeam_channel::Receiver<WebXRLayerManagerFactory<WebXRSurfman>>,
+//     managers: HashMap<WebXRLayerManagerId, Box<dyn WebXRLayerManagerAPI<WebXRSurfman>>>,
+//     next_manager_id: u32,
+// }
 
-#[cfg(feature = "webxr")]
-impl WebXRBridge {
-    pub(crate) fn new(init: WebXRBridgeInit) -> WebXRBridge {
-        let WebXRBridgeInit {
-            factory_receiver, ..
-        } = init;
-        let managers = HashMap::new();
-        let next_manager_id = 1;
-        WebXRBridge {
-            factory_receiver,
-            managers,
-            next_manager_id,
-        }
-    }
-}
+// #[cfg(feature = "webxr")]
+// impl WebXRBridge {
+//     pub(crate) fn new(init: WebXRBridgeInit) -> WebXRBridge {
+//         let WebXRBridgeInit {
+//             factory_receiver, ..
+//         } = init;
+//         let managers = HashMap::new();
+//         let next_manager_id = 1;
+//         WebXRBridge {
+//             factory_receiver,
+//             managers,
+//             next_manager_id,
+//         }
+//     }
+// }
 
-#[cfg(feature = "webxr")]
-impl WebXRBridge {
-    #[allow(unsafe_code)]
-    fn create_layer_manager(
-        &mut self,
-        device: &mut Device,
-        contexts: &mut dyn WebXRContexts<WebXRSurfman>,
-    ) -> Result<WebXRLayerManagerId, WebXRError> {
-        let factory = self
-            .factory_receiver
-            .recv()
-            .map_err(|_| WebXRError::CommunicationError)?;
-        let manager = factory.build(device, contexts)?;
-        let manager_id = unsafe { WebXRLayerManagerId::new(self.next_manager_id) };
-        self.next_manager_id += 1;
-        self.managers.insert(manager_id, manager);
-        Ok(manager_id)
-    }
+// #[cfg(feature = "webxr")]
+// impl WebXRBridge {
+//     #[allow(unsafe_code)]
+//     fn create_layer_manager(
+//         &mut self,
+//         device: &mut Device,
+//         contexts: &mut dyn WebXRContexts<WebXRSurfman>,
+//     ) -> Result<WebXRLayerManagerId, WebXRError> {
+//         let factory = self
+//             .factory_receiver
+//             .recv()
+//             .map_err(|_| WebXRError::CommunicationError)?;
+//         let manager = factory.build(device, contexts)?;
+//         let manager_id = unsafe { WebXRLayerManagerId::new(self.next_manager_id) };
+//         self.next_manager_id += 1;
+//         self.managers.insert(manager_id, manager);
+//         Ok(manager_id)
+//     }
 
-    fn destroy_layer_manager(&mut self, manager_id: WebXRLayerManagerId) {
-        self.managers.remove(&manager_id);
-    }
+//     fn destroy_layer_manager(&mut self, manager_id: WebXRLayerManagerId) {
+//         self.managers.remove(&manager_id);
+//     }
 
-    fn create_layer(
-        &mut self,
-        manager_id: WebXRLayerManagerId,
-        device: &mut Device,
-        contexts: &mut dyn WebXRContexts<WebXRSurfman>,
-        context_id: WebXRContextId,
-        layer_init: WebXRLayerInit,
-    ) -> Result<WebXRLayerId, WebXRError> {
-        let manager = self
-            .managers
-            .get_mut(&manager_id)
-            .ok_or(WebXRError::NoMatchingDevice)?;
-        manager.create_layer(device, contexts, context_id, layer_init)
-    }
+//     fn create_layer(
+//         &mut self,
+//         manager_id: WebXRLayerManagerId,
+//         device: &mut Device,
+//         contexts: &mut dyn WebXRContexts<WebXRSurfman>,
+//         context_id: WebXRContextId,
+//         layer_init: WebXRLayerInit,
+//     ) -> Result<WebXRLayerId, WebXRError> {
+//         let manager = self
+//             .managers
+//             .get_mut(&manager_id)
+//             .ok_or(WebXRError::NoMatchingDevice)?;
+//         manager.create_layer(device, contexts, context_id, layer_init)
+//     }
 
-    fn destroy_layer(
-        &mut self,
-        manager_id: WebXRLayerManagerId,
-        device: &mut Device,
-        contexts: &mut dyn WebXRContexts<WebXRSurfman>,
-        context_id: WebXRContextId,
-        layer_id: WebXRLayerId,
-    ) {
-        if let Some(manager) = self.managers.get_mut(&manager_id) {
-            manager.destroy_layer(device, contexts, context_id, layer_id);
-        }
-    }
+//     fn destroy_layer(
+//         &mut self,
+//         manager_id: WebXRLayerManagerId,
+//         device: &mut Device,
+//         contexts: &mut dyn WebXRContexts<WebXRSurfman>,
+//         context_id: WebXRContextId,
+//         layer_id: WebXRLayerId,
+//     ) {
+//         if let Some(manager) = self.managers.get_mut(&manager_id) {
+//             manager.destroy_layer(device, contexts, context_id, layer_id);
+//         }
+//     }
 
-    fn destroy_all_layers(
-        &mut self,
-        device: &mut Device,
-        contexts: &mut dyn WebXRContexts<WebXRSurfman>,
-        context_id: WebXRContextId,
-    ) {
-        for manager in self.managers.values_mut() {
-            #[allow(clippy::unnecessary_to_owned)] // Needs mutable borrow later in destroy
-            for (other_id, layer_id) in manager.layers().to_vec() {
-                if other_id == context_id {
-                    manager.destroy_layer(device, contexts, context_id, layer_id);
-                }
-            }
-        }
-    }
+//     fn destroy_all_layers(
+//         &mut self,
+//         device: &mut Device,
+//         contexts: &mut dyn WebXRContexts<WebXRSurfman>,
+//         context_id: WebXRContextId,
+//     ) {
+//         for manager in self.managers.values_mut() {
+//             #[allow(clippy::unnecessary_to_owned)] // Needs mutable borrow later in destroy
+//             for (other_id, layer_id) in manager.layers().to_vec() {
+//                 if other_id == context_id {
+//                     manager.destroy_layer(device, contexts, context_id, layer_id);
+//                 }
+//             }
+//         }
+//     }
 
-    fn begin_frame(
-        &mut self,
-        manager_id: WebXRLayerManagerId,
-        device: &mut Device,
-        contexts: &mut dyn WebXRContexts<WebXRSurfman>,
-        layers: &[(WebXRContextId, WebXRLayerId)],
-    ) -> Result<Vec<WebXRSubImages>, WebXRError> {
-        let manager = self
-            .managers
-            .get_mut(&manager_id)
-            .ok_or(WebXRError::NoMatchingDevice)?;
-        manager.begin_frame(device, contexts, layers)
-    }
+//     fn begin_frame(
+//         &mut self,
+//         manager_id: WebXRLayerManagerId,
+//         device: &mut Device,
+//         contexts: &mut dyn WebXRContexts<WebXRSurfman>,
+//         layers: &[(WebXRContextId, WebXRLayerId)],
+//     ) -> Result<Vec<WebXRSubImages>, WebXRError> {
+//         let manager = self
+//             .managers
+//             .get_mut(&manager_id)
+//             .ok_or(WebXRError::NoMatchingDevice)?;
+//         manager.begin_frame(device, contexts, layers)
+//     }
 
-    fn end_frame(
-        &mut self,
-        manager_id: WebXRLayerManagerId,
-        device: &mut Device,
-        contexts: &mut dyn WebXRContexts<WebXRSurfman>,
-        layers: &[(WebXRContextId, WebXRLayerId)],
-    ) -> Result<(), WebXRError> {
-        let manager = self
-            .managers
-            .get_mut(&manager_id)
-            .ok_or(WebXRError::NoMatchingDevice)?;
-        manager.end_frame(device, contexts, layers)
-    }
-}
+//     fn end_frame(
+//         &mut self,
+//         manager_id: WebXRLayerManagerId,
+//         device: &mut Device,
+//         contexts: &mut dyn WebXRContexts<WebXRSurfman>,
+//         layers: &[(WebXRContextId, WebXRLayerId)],
+//     ) -> Result<(), WebXRError> {
+//         let manager = self
+//             .managers
+//             .get_mut(&manager_id)
+//             .ok_or(WebXRError::NoMatchingDevice)?;
+//         manager.end_frame(device, contexts, layers)
+//     }
+// }
 
-#[cfg(feature = "webxr")]
-pub(crate) struct WebXRBridgeInit {
-    sender: WebGLSender<WebGLMsg>,
-    factory_receiver: crossbeam_channel::Receiver<WebXRLayerManagerFactory<WebXRSurfman>>,
-    factory_sender: crossbeam_channel::Sender<WebXRLayerManagerFactory<WebXRSurfman>>,
-}
+// #[cfg(feature = "webxr")]
+// pub(crate) struct WebXRBridgeInit {
+//     sender: WebGLSender<WebGLMsg>,
+//     factory_receiver: crossbeam_channel::Receiver<WebXRLayerManagerFactory<WebXRSurfman>>,
+//     factory_sender: crossbeam_channel::Sender<WebXRLayerManagerFactory<WebXRSurfman>>,
+// }
 
-#[cfg(feature = "webxr")]
-impl WebXRBridgeInit {
-    pub(crate) fn new(sender: WebGLSender<WebGLMsg>) -> WebXRBridgeInit {
-        let (factory_sender, factory_receiver) = crossbeam_channel::unbounded();
-        WebXRBridgeInit {
-            sender,
-            factory_sender,
-            factory_receiver,
-        }
-    }
+// #[cfg(feature = "webxr")]
+// impl WebXRBridgeInit {
+//     pub(crate) fn new(sender: WebGLSender<WebGLMsg>) -> WebXRBridgeInit {
+//         let (factory_sender, factory_receiver) = crossbeam_channel::unbounded();
+//         WebXRBridgeInit {
+//             sender,
+//             factory_sender,
+//             factory_receiver,
+//         }
+//     }
 
-    pub(crate) fn layer_grand_manager(&self) -> WebXRLayerGrandManager<WebXRSurfman> {
-        WebXRLayerGrandManager::new(WebXRBridgeGrandManager {
-            sender: self.sender.clone(),
-            factory_sender: self.factory_sender.clone(),
-        })
-    }
-}
+//     pub(crate) fn layer_grand_manager(&self) -> WebXRLayerGrandManager<WebXRSurfman> {
+//         WebXRLayerGrandManager::new(WebXRBridgeGrandManager {
+//             sender: self.sender.clone(),
+//             factory_sender: self.factory_sender.clone(),
+//         })
+//     }
+// }
 
-#[cfg(feature = "webxr")]
-struct WebXRBridgeGrandManager {
-    sender: WebGLSender<WebGLMsg>,
-    // WebXR layer manager factories use generic trait objects under the
-    // hood, which aren't deserializable (even using typetag)
-    // so we can't send them over the regular webgl channel.
-    // Fortunately, the webgl thread runs in the same process as
-    // the webxr threads, so we can use a crossbeam channel to send
-    // factories.
-    factory_sender: crossbeam_channel::Sender<WebXRLayerManagerFactory<WebXRSurfman>>,
-}
+// #[cfg(feature = "webxr")]
+// struct WebXRBridgeGrandManager {
+//     sender: WebGLSender<WebGLMsg>,
+//     // WebXR layer manager factories use generic trait objects under the
+//     // hood, which aren't deserializable (even using typetag)
+//     // so we can't send them over the regular webgl channel.
+//     // Fortunately, the webgl thread runs in the same process as
+//     // the webxr threads, so we can use a crossbeam channel to send
+//     // factories.
+//     factory_sender: crossbeam_channel::Sender<WebXRLayerManagerFactory<WebXRSurfman>>,
+// }
 
-#[cfg(feature = "webxr")]
-impl WebXRLayerGrandManagerAPI<WebXRSurfman> for WebXRBridgeGrandManager {
-    fn create_layer_manager(
-        &self,
-        factory: WebXRLayerManagerFactory<WebXRSurfman>,
-    ) -> Result<WebXRLayerManager, WebXRError> {
-        let (sender, receiver) = webgl_channel().ok_or(WebXRError::CommunicationError)?;
-        let _ = self.factory_sender.send(factory);
-        let _ = self
-            .sender
-            .send(WebGLMsg::WebXRCommand(WebXRCommand::CreateLayerManager(
-                sender,
-            )));
-        let sender = self.sender.clone();
-        let manager_id = receiver
-            .recv()
-            .map_err(|_| WebXRError::CommunicationError)??;
-        let layers = Vec::new();
-        Ok(WebXRLayerManager::new(WebXRBridgeManager {
-            manager_id,
-            sender,
-            layers,
-        }))
-    }
+// #[cfg(feature = "webxr")]
+// impl WebXRLayerGrandManagerAPI<WebXRSurfman> for WebXRBridgeGrandManager {
+//     fn create_layer_manager(
+//         &self,
+//         factory: WebXRLayerManagerFactory<WebXRSurfman>,
+//     ) -> Result<WebXRLayerManager, WebXRError> {
+//         let (sender, receiver) = webgl_channel().ok_or(WebXRError::CommunicationError)?;
+//         let _ = self.factory_sender.send(factory);
+//         let _ = self
+//             .sender
+//             .send(WebGLMsg::WebXRCommand(WebXRCommand::CreateLayerManager(
+//                 sender,
+//             )));
+//         let sender = self.sender.clone();
+//         let manager_id = receiver
+//             .recv()
+//             .map_err(|_| WebXRError::CommunicationError)??;
+//         let layers = Vec::new();
+//         Ok(WebXRLayerManager::new(WebXRBridgeManager {
+//             manager_id,
+//             sender,
+//             layers,
+//         }))
+//     }
 
-    fn clone_layer_grand_manager(&self) -> WebXRLayerGrandManager<WebXRSurfman> {
-        WebXRLayerGrandManager::new(WebXRBridgeGrandManager {
-            sender: self.sender.clone(),
-            factory_sender: self.factory_sender.clone(),
-        })
-    }
-}
+//     fn clone_layer_grand_manager(&self) -> WebXRLayerGrandManager<WebXRSurfman> {
+//         WebXRLayerGrandManager::new(WebXRBridgeGrandManager {
+//             sender: self.sender.clone(),
+//             factory_sender: self.factory_sender.clone(),
+//         })
+//     }
+// }
 
-#[cfg(feature = "webxr")]
-struct WebXRBridgeManager {
-    sender: WebGLSender<WebGLMsg>,
-    manager_id: WebXRLayerManagerId,
-    layers: Vec<(WebXRContextId, WebXRLayerId)>,
-}
+// // #[cfg(feature = "webxr")]
+// // struct WebXRBridgeManager {
+// //     sender: WebGLSender<WebGLMsg>,
+// //     manager_id: WebXRLayerManagerId,
+// //     layers: Vec<(WebXRContextId, WebXRLayerId)>,
+// // }
 
-#[cfg(feature = "webxr")]
-impl<GL: WebXRTypes> WebXRLayerManagerAPI<GL> for WebXRBridgeManager {
-    fn create_layer(
-        &mut self,
-        _: &mut GL::Device,
-        _: &mut dyn WebXRContexts<GL>,
-        context_id: WebXRContextId,
-        init: WebXRLayerInit,
-    ) -> Result<WebXRLayerId, WebXRError> {
-        let (sender, receiver) = webgl_channel().ok_or(WebXRError::CommunicationError)?;
-        let _ = self
-            .sender
-            .send(WebGLMsg::WebXRCommand(WebXRCommand::CreateLayer(
-                self.manager_id,
-                context_id,
-                init,
-                sender,
-            )));
-        let layer_id = receiver
-            .recv()
-            .map_err(|_| WebXRError::CommunicationError)??;
-        self.layers.push((context_id, layer_id));
-        Ok(layer_id)
-    }
+// #[cfg(feature = "webxr")]
+// impl<GL: WebXRTypes> WebXRLayerManagerAPI<GL> for WebXRBridgeManager {
+//     fn create_layer(
+//         &mut self,
+//         _: &mut GL::Device,
+//         _: &mut dyn WebXRContexts<GL>,
+//         context_id: WebXRContextId,
+//         init: WebXRLayerInit,
+//     ) -> Result<WebXRLayerId, WebXRError> {
+//         let (sender, receiver) = webgl_channel().ok_or(WebXRError::CommunicationError)?;
+//         let _ = self
+//             .sender
+//             .send(WebGLMsg::WebXRCommand(WebXRCommand::CreateLayer(
+//                 self.manager_id,
+//                 context_id,
+//                 init,
+//                 sender,
+//             )));
+//         let layer_id = receiver
+//             .recv()
+//             .map_err(|_| WebXRError::CommunicationError)??;
+//         self.layers.push((context_id, layer_id));
+//         Ok(layer_id)
+//     }
 
-    fn destroy_layer(
-        &mut self,
-        _: &mut GL::Device,
-        _: &mut dyn WebXRContexts<GL>,
-        context_id: WebXRContextId,
-        layer_id: WebXRLayerId,
-    ) {
-        self.layers.retain(|&ids| ids != (context_id, layer_id));
-        let _ = self
-            .sender
-            .send(WebGLMsg::WebXRCommand(WebXRCommand::DestroyLayer(
-                self.manager_id,
-                context_id,
-                layer_id,
-            )));
-    }
+//     fn destroy_layer(
+//         &mut self,
+//         _: &mut GL::Device,
+//         _: &mut dyn WebXRContexts<GL>,
+//         context_id: WebXRContextId,
+//         layer_id: WebXRLayerId,
+//     ) {
+//         self.layers.retain(|&ids| ids != (context_id, layer_id));
+//         let _ = self
+//             .sender
+//             .send(WebGLMsg::WebXRCommand(WebXRCommand::DestroyLayer(
+//                 self.manager_id,
+//                 context_id,
+//                 layer_id,
+//             )));
+//     }
 
-    fn layers(&self) -> &[(WebXRContextId, WebXRLayerId)] {
-        &self.layers[..]
-    }
+//     fn layers(&self) -> &[(WebXRContextId, WebXRLayerId)] {
+//         &self.layers[..]
+//     }
 
-    fn begin_frame(
-        &mut self,
-        _: &mut GL::Device,
-        _: &mut dyn WebXRContexts<GL>,
-        layers: &[(WebXRContextId, WebXRLayerId)],
-    ) -> Result<Vec<WebXRSubImages>, WebXRError> {
-        let (sender, receiver) = webgl_channel().ok_or(WebXRError::CommunicationError)?;
-        let _ = self
-            .sender
-            .send(WebGLMsg::WebXRCommand(WebXRCommand::BeginFrame(
-                self.manager_id,
-                layers.to_vec(),
-                sender,
-            )));
-        receiver
-            .recv()
-            .map_err(|_| WebXRError::CommunicationError)?
-    }
+//     fn begin_frame(
+//         &mut self,
+//         _: &mut GL::Device,
+//         _: &mut dyn WebXRContexts<GL>,
+//         layers: &[(WebXRContextId, WebXRLayerId)],
+//     ) -> Result<Vec<WebXRSubImages>, WebXRError> {
+//         let (sender, receiver) = webgl_channel().ok_or(WebXRError::CommunicationError)?;
+//         let _ = self
+//             .sender
+//             .send(WebGLMsg::WebXRCommand(WebXRCommand::BeginFrame(
+//                 self.manager_id,
+//                 layers.to_vec(),
+//                 sender,
+//             )));
+//         receiver
+//             .recv()
+//             .map_err(|_| WebXRError::CommunicationError)?
+//     }
 
-    fn end_frame(
-        &mut self,
-        _: &mut GL::Device,
-        _: &mut dyn WebXRContexts<GL>,
-        layers: &[(WebXRContextId, WebXRLayerId)],
-    ) -> Result<(), WebXRError> {
-        let (sender, receiver) = webgl_channel().ok_or(WebXRError::CommunicationError)?;
-        let _ = self
-            .sender
-            .send(WebGLMsg::WebXRCommand(WebXRCommand::EndFrame(
-                self.manager_id,
-                layers.to_vec(),
-                sender,
-            )));
-        receiver
-            .recv()
-            .map_err(|_| WebXRError::CommunicationError)?
-    }
-}
+//     fn end_frame(
+//         &mut self,
+//         _: &mut GL::Device,
+//         _: &mut dyn WebXRContexts<GL>,
+//         layers: &[(WebXRContextId, WebXRLayerId)],
+//     ) -> Result<(), WebXRError> {
+//         let (sender, receiver) = webgl_channel().ok_or(WebXRError::CommunicationError)?;
+//         let _ = self
+//             .sender
+//             .send(WebGLMsg::WebXRCommand(WebXRCommand::EndFrame(
+//                 self.manager_id,
+//                 layers.to_vec(),
+//                 sender,
+//             )));
+//         receiver
+//             .recv()
+//             .map_err(|_| WebXRError::CommunicationError)?
+//     }
+// }
 
-#[cfg(feature = "webxr")]
-impl Drop for WebXRBridgeManager {
-    fn drop(&mut self) {
-        let _ = self
-            .sender
-            .send(WebGLMsg::WebXRCommand(WebXRCommand::DestroyLayerManager(
-                self.manager_id,
-            )));
-    }
-}
+// #[cfg(feature = "webxr")]
+// impl Drop for WebXRBridgeManager {
+//     fn drop(&mut self) {
+//         let _ = self
+//             .sender
+//             .send(WebGLMsg::WebXRCommand(WebXRCommand::DestroyLayerManager(
+//                 self.manager_id,
+//             )));
+//     }
+// }
 
-#[cfg(feature = "webxr")]
-struct WebXRBridgeContexts<'a> {
-    contexts: &'a mut FnvHashMap<WebGLContextId, GLContextData>,
-    bound_context_id: &'a mut Option<WebGLContextId>,
-}
+// #[cfg(feature = "webxr")]
+// struct WebXRBridgeContexts<'a> {
+//     contexts: &'a mut FnvHashMap<WebGLContextId, GLContextData>,
+//     bound_context_id: &'a mut Option<WebGLContextId>,
+// }
 
-#[cfg(feature = "webxr")]
-impl<'a> WebXRContexts<WebXRSurfman> for WebXRBridgeContexts<'a> {
-    fn context(&mut self, device: &Device, context_id: WebXRContextId) -> Option<&mut Context> {
-        let data = WebGLThread::make_current_if_needed_mut(
-            device,
-            WebGLContextId::from(context_id),
-            self.contexts,
-            self.bound_context_id,
-        )?;
-        Some(&mut data.ctx)
-    }
-    fn bindings(&mut self, device: &Device, context_id: WebXRContextId) -> Option<&Gl> {
-        let data = WebGLThread::make_current_if_needed(
-            device,
-            WebGLContextId::from(context_id),
-            self.contexts,
-            self.bound_context_id,
-        )?;
-        Some(&data.gl)
-    }
-}
+// #[cfg(feature = "webxr")]
+// impl<'a> WebXRContexts<WebXRSurfman> for WebXRBridgeContexts<'a> {
+//     fn context(&mut self, device: &Device, context_id: WebXRContextId) -> Option<&mut Context> {
+//         let data = WebGLThread::make_current_if_needed_mut(
+//             device,
+//             WebGLContextId::from(context_id),
+//             self.contexts,
+//             self.bound_context_id,
+//         )?;
+//         Some(&mut data.ctx)
+//     }
+//     fn bindings(&mut self, device: &Device, context_id: WebXRContextId) -> Option<&Gl> {
+//         let data = WebGLThread::make_current_if_needed(
+//             device,
+//             WebGLContextId::from(context_id),
+//             self.contexts,
+//             self.bound_context_id,
+//         )?;
+//         Some(&data.gl)
+//     }
+// }
